header, footer, notice, faq css 수정

+++ src/main/webapp/WEB-INF/views/admin/faq/list.jsp
@@ -25,17 +25,24 @@ function submitContents2(elClickedObj) {
 $(document).ready(function(){
 	//버튼 클릭에 따른 동작
 	$("#btnWriteProc").click(function(){
-		submitContents($("#btnWriteProc"));
-		$("#formWrite").submit();
+		if(confirm("새로운 게시글을 등록하시겠습니까?")){
+			submitContents($("#btnWriteProc"));
+			$("#formWrite").submit();
+		}else{
+			return false;
+		}
 	})
 	
 	//버튼 클릭에 따른 동작
 	$("#btnUpdateProc").click(function(){
-		submitContents2($("#btnUpdateProc"));
-		$("#formUpdate").submit();
+		if(confirm("게시글을 수정하시겠습니까? 수정된 게시물은 원상복구 할 수 없습니다.")){
+			submitContents2($("#btnUpdateProc"));
+			$("#formUpdate").submit();			
+		}else{
+			return false;
+		}
 	})
 	
-
 	//게시글 상세보기 토글
     $(".titleChild").click(function(){
 		$(this).parent().next("div").children().slideToggle();
@@ -123,14 +130,16 @@ $(document).ready(function(){
 }
 
 .row{
-	width: 1100px;
+	width: 900px;
+	margin: 0 auto;
 }
 .title{
 	background: #8A7E6B;
-	color: #ECE6CC;
+	color: #35312B;
 	font-size: 21px;
 	font-weight: 700;
  	margin-bottom: 5px;
+ 	border-radius: 10px;
 }
 .titleText{
 	margin: 20px;
@@ -138,33 +147,50 @@ $(document).ready(function(){
 .contentChild{
 	background: #ECE6CC;
 	margin-bottom: 5px;
+	padding: 30px;
+ 	border-radius: 10px;
 /* 	overflow-y: scroll; */
 }
 .title button{
 	color: pink;
 }
 .title a:link{
+	vertical-align: middle;
 	color: #ECE6CC;
 }
 .title a:hover{
-	color: #35312B;
+	color: #8A7E6B;
+	text-decoration: none;
+}
+.titleChild:hover{
+	background-color: #35312B;
+	color: #ECE6CC;
+	border-radius: 10px;
+}
+.titleChild{
+	padding-right: 15px;
+	padding-left: 15px;
+}
+.btnWrite{
+	background-color: #8A7E6B;
+	color: #ECE6CC;
+}
+.btnWrite:hover{
+	background-color: #35312B;
+	color: #ECE6CC;
 }
 </style>
 
 
 <div class="container">
 
-<h1>admin faq list + write &nbsp;&nbsp;&nbsp;
-<a href="/admin/faq/write"><button id="btnWrite" class="btn">WRITE</button></a>
-<button id="modalWrite" class="btn btn-open-popup-write">모달 글쓰기</button></h1>
-총 게시글 수 : [${faqTotal }] (>= 10이면 글쓰기 버튼 비활성화 또는 알럹)
-
-
+<h1>admin faq &nbsp;&nbsp;&nbsp;
+<button id="modalWrite" class="btn btn-open-popup-write">글쓰기</button></h1>
 <hr>
 
 <%-- 모달 글쓰기 영역 --%>
 <div class="modalWrite">
-	<div class="modalWrite_body">새로운 글 등록하기
+	<div class="modalWrite_body">새로운 자주 묻는 문의 등록하기
 		<form id="formWrite" action="/admin/faq/write" method="post">
 			<label for="title">제목</label>
 			<input type="text" id="title" name="faqTitle" class="form-control"/>
@@ -172,7 +198,7 @@ $(document).ready(function(){
 			<label for="contentWrite">본문</label>
 			<textarea rows="10" style="width: 100%;" id="contentWrite" name="faqContent"></textarea>
 		</form>
-		<button class="btn" id="btnWriteProc">작성</button>
+		<button class="btn btnWrite" id="btnWriteProc">작성</button>
 	</div>
 </div>
 
@@ -180,11 +206,9 @@ $(document).ready(function(){
 <div class="row" data-boardno="${i.faqNo }">
 	<div class="title">
 		<div class="titleChild" style="cursor: pointer;">
-		<span class="titleText">${i.faqTitle}</span>
-		<span style="float: right; font-size: 18px; font-weight: 400;"><a href="/admin/faq/delete?faqNo=${i.faqNo}">삭제</a>&nbsp;&nbsp;&nbsp;</span>
-		<span style="float: right; font-size: 18px; font-weight: 400;"><a href="#" class="btn-open-popup-update">수정</a>&nbsp;&nbsp;&nbsp;</span>
-<%-- 		<span class="pull-right"><button class="btn pull-right" id="btnDelete" onclick="btnDelete(${i.faqNo})">삭제</button>&nbsp;&nbsp;&nbsp;</span> --%>
-<%-- 		<span class="pull-right"><button class="btn btn-open-popup-update pull-right">수정</button>&nbsp;&nbsp;&nbsp;</span> --%>
+			<span class="titleText">${i.faqTitle}</span>
+			<span style="float: right; font-size: 18px; font-weight: 400;"><a href="/admin/faq/delete?faqNo=${i.faqNo}" onclick="return confirm('게시글을 삭제하시겠습니까? 삭제된 게시물은 원상복구 할 수 없습니다.');">삭제</a>&nbsp;&nbsp;&nbsp;</span>
+			<span style="float: right; font-size: 18px; font-weight: 400;"><a href="#" class="btn-open-popup-update">수정</a>&nbsp;&nbsp;&nbsp;</span>
 		</div>
 	</div>
 	<div class="content">
@@ -192,18 +216,14 @@ $(document).ready(function(){
 			<div class="contentText">
 			${i.faqContent}
 			</div>
-			<div class="contentAnchor">
-			내가 찾는 내용이 아니에요 
-			<a href="#"><button>채팅 문의로 이동하기</button></a>
-			</div>
 		</div>
 	</div>
 </div>
 </c:forEach>
 
-<%-- 수정을 위해 모달로 띄울 영역 --%>
+<%-- 모달 수정 영역 --%>
 <div id="modalUpdate" >
-	<div class="modal_body">모달 수정 영역
+	<div class="modal_body">자주 묻는 문의 수정하기
 		<form id="formUpdate" action="/admin/faq/update" method="post">
 		<input type="hidden" name="faqNo" id="boardNo" />
 		
@@ -213,7 +233,7 @@ $(document).ready(function(){
 			<label for="contentUpdtae">본문</label>
 			<textarea rows="10" style="width: 100%;" id="contentUpdate" name="faqContent"></textarea>
 		</form>
-		<button class="btn" id="btnUpdateProc">작성</button>
+		<button class="btn btnWrite" id="btnUpdateProc">작성</button>
 	</div>
 </div>
 
@@ -244,16 +264,20 @@ const modalWrite = document.querySelector('.modalWrite');
 const btnOpenPopupWrite = document.querySelector('.btn-open-popup-write');
 
 btnOpenPopupWrite.addEventListener('click', () => {
-	$(".modalWrite iframe").remove()
-	
-	//btnOpenPopupWrite에 맞춰서 SE2 스킨 씌우기
-	initSEWrite();
-	
-	modalWrite.classList.toggle('show');
-	
-    if (modalWrite.classList.contains('show')) {
-    	bodyWrite.style.overflow = 'hidden';
-    }
+	if(${faqTotal }>=10){
+		alert("자주 묻는 문의는 10개를 초과하여 등록할 수 없습니다.")
+	}else{
+		$(".modalWrite iframe").remove()
+		
+		//btnOpenPopupWrite에 맞춰서 SE2 스킨 씌우기
+		initSEWrite();
+		
+		modalWrite.classList.toggle('show');
+		
+	    if (modalWrite.classList.contains('show')) {
+	    	bodyWrite.style.overflow = 'hidden';
+	    }		
+	}
 });
 
 modalWrite.addEventListener('click', (event) => {
